+++ src/main/java/com/alternacraft/pvptitles/Hook/SBSHook.java
@@ -65,7 +65,7 @@ public class SBSHook {
                     LoggerManager.logError(ex.getCustomMessage());
                 }
 
-                int seconds = 0;
+                long seconds = 0;
                 try {
                     seconds = plugin.getManager().dbh.getDm().loadPlayedTime(player.getUniqueId());
                 } catch (DBException ex) {
